move: ModelOfferPage pages 폴더로 이동
