+++ src/App.tsx
@@ -12,6 +12,7 @@ import ErrorPage from './pages/ErrorPage';
 import LoginPage from './pages/LoginPage';
 import MainPage from './pages/MainPage';
 import ModelInfoPage from './pages/ModelInfoPage';
+import ModelOfferPage from './pages/ModelOfferPage';
 import MyPage from './pages/MyPage';
 import MyQuitPage from './pages/MyQuitPage';
 import OfferInfoPage from './pages/OfferInfoPage';
@@ -19,7 +20,6 @@ import SignUpPage from './pages/SignUpPage';
 import GlobalStyle from './styles/GlobalStyle';
 import theme from './styles/Theme';
 import LoginCallback from './views/LoginPage/components/LoginCallback';
-import ModelOfferPage from './views/ModelInfoPage/components/ModelOfferPage';
 import OfferSentCompletePage from './views/ModelInfoPage/components/OfferSentCompletePage';
 import CheckOfferPage from './views/OfferInfoPage/components/CheckOfferPage';
